+++ packages/prism/lib/palettes/solarized.dart
@@ -1,0 +1,141 @@
+// GENERATED CODE - DO NOT EDIT BY HAND
+// ignore_for_file: public_member_api_docs
+import 'package:prism/prism.dart';
+
+/// An enhanced enum for the SolarizedPalette palette.
+enum SolarizedPalette {
+  base03(RayScheme(
+    ray: Ray(0xFF002B36),
+    luminance: 0.019940957046130022,
+    isDark: true,
+    onRay: Ray(0xFFFFFFFF),
+    surfaceDark: Ray(0xFF153C46),
+    surfaceLight: Ray(0xFFE5E9EA),
+  )),
+  base02(RayScheme(
+    ray: Ray(0xFF073642),
+    luminance: 0.030768519742928466,
+    isDark: true,
+    onRay: Ray(0xFFFFFFFF),
+    surfaceDark: Ray(0xFF194450),
+    surfaceLight: Ray(0xFFE5EAEC),
+  )),
+  base01(RayScheme(
+    ray: Ray(0xFF586E75),
+    luminance: 0.14510922048021613,
+    isDark: true,
+    onRay: Ray(0xFFFFFFFF),
+    surfaceDark: Ray(0xFF3D4C51),
+    surfaceLight: Ray(0xFFEBEEEF),
+  )),
+  base00(RayScheme(
+    ray: Ray(0xFF657B83),
+    luminance: 0.18571312836107598,
+    isDark: true,
+    onRay: Ray(0xFFFFFFFF),
+    surfaceDark: Ray(0xFF364247),
+    surfaceLight: Ray(0xFFECEFF0),
+  )),
+  base0(RayScheme(
+    ray: Ray(0xFF839496),
+    luminance: 0.28207112005905194,
+    isDark: true,
+    onRay: Ray(0xFFFFFFFF),
+    surfaceDark: Ray(0xFF2E3435),
+    surfaceLight: Ray(0xFFEEF0F0),
+  )),
+  base1(RayScheme(
+    ray: Ray(0xFF93A1A1),
+    luminance: 0.342659913651724,
+    isDark: true,
+    onRay: Ray(0xFFFFFFFF),
+    surfaceDark: Ray(0xFF2B2F2F),
+    surfaceLight: Ray(0xFFEFF1F1),
+  )),
+  base2(RayScheme(
+    ray: Ray(0xFFEEE8D5),
+    luminance: 0.8069446461082409,
+    isDark: false,
+    onRay: Ray(0xFF000000),
+    surfaceDark: Ray(0xFF1D1D1A),
+    surfaceLight: Ray(0xFFF6F3E9),
+  )),
+  base3(RayScheme(
+    ray: Ray(0xFFFDF6E3),
+    luminance: 0.9234023234416118,
+    isDark: false,
+    onRay: Ray(0xFF000000),
+    surfaceDark: Ray(0xFF1B1B19),
+    surfaceLight: Ray(0xFFF7F0DD),
+  )),
+  yellow(RayScheme(
+    ray: Ray(0xFFB58900),
+    luminance: 0.2771507753653897,
+    isDark: true,
+    onRay: Ray(0xFFFFFFFF),
+    surfaceDark: Ray(0xFF413100),
+    surfaceLight: Ray(0xFFF5EFDC),
+  )),
+  orange(RayScheme(
+    ray: Ray(0xFFCB4B16),
+    luminance: 0.1778659151818863,
+    isDark: true,
+    onRay: Ray(0xFFFFFFFF),
+    surfaceDark: Ray(0xFF722A0C),
+    surfaceLight: Ray(0xFFF9E9E3),
+  )),
+  red(RayScheme(
+    ray: Ray(0xFFDC322F),
+    luminance: 0.17702084111371486,
+    isDark: true,
+    onRay: Ray(0xFFFFFFFF),
+    surfaceDark: Ray(0xFF7C1C1B),
+    surfaceLight: Ray(0xFFFBE6E6),
+  )),
+  magenta(RayScheme(
+    ray: Ray(0xFFD33682),
+    luminance: 0.18098923196059497,
+    isDark: true,
+    onRay: Ray(0xFFFFFFFF),
+    surfaceDark: Ray(0xFF751E48),
+    surfaceLight: Ray(0xFFFAE6F0),
+  )),
+  violet(RayScheme(
+    ray: Ray(0xFF6C71C4),
+    luminance: 0.18983922001049444,
+    isDark: true,
+    onRay: Ray(0xFFFFFFFF),
+    surfaceDark: Ray(0xFF393C67),
+    surfaceLight: Ray(0xFFEDEDF8),
+  )),
+  blue(RayScheme(
+    ray: Ray(0xFF268BD2),
+    luminance: 0.23530449938724587,
+    isDark: true,
+    onRay: Ray(0xFFFFFFFF),
+    surfaceDark: Ray(0xFF103B59),
+    surfaceLight: Ray(0xFFE3F0F9),
+  )),
+  cyan(RayScheme(
+    ray: Ray(0xFF2AA198),
+    luminance: 0.2824897738483719,
+    isDark: true,
+    onRay: Ray(0xFFFFFFFF),
+    surfaceDark: Ray(0xFF0F3936),
+    surfaceLight: Ray(0xFFE1F2F1),
+  )),
+  green(RayScheme(
+    ray: Ray(0xFF859900),
+    luminance: 0.2776901094824954,
+    isDark: true,
+    onRay: Ray(0xFFFFFFFF),
+    surfaceDark: Ray(0xFF303700),
+    surfaceLight: Ray(0xFFEEF1DC),
+  )),
+;
+
+  /// The pre-computed profile for this color.
+  final RayScheme scheme;
+
+  const SolarizedPalette(this.scheme);
+}