+++ packages/prism/lib/prism.dart
@@ -1,3 +1,2 @@
-library prism;
-
 export 'src/ray.dart';
+export 'src/ray_scheme.dart';